Move most of 'tracking.tools' to new 'streamline_operations'
